Get the entry assembly instead of the API DLL.
Get the entry assembly instead of the API DLL.
Fix possible buffer overflow.
Username should be optional on Run method.
Fix uninstall service from last commit.

+++ IOTLinkAPI/Platform/Windows/WindowsAPI.cs
@@ -317,7 +317,7 @@ namespace IOTLink.Platform.Windows
             int count = 0;
             int retval = WtsApi32.WTSEnumerateSessions(server, 0, 1, ref buffer, ref count);
             int dataSize = Marshal.SizeOf(typeof(WtsApi32.WtsSessionInfo));
-            Int64 current = (int)buffer;
+            Int64 current = (long)buffer;
 
             if (retval != 0)
             {
